+++ src/main/java/org/example/dto/LandmarkSaveRequestDTO.java
@@ -21,7 +21,7 @@ public class LandmarkSaveRequestDTO {
         private String image;
         private String open;
         private String close;
-        private Double lat;
-        private Double lon;
+        private Float lat;
+        private Float lon;
     }
