+++ src/main/java/org/example/dto/LandmarkSaveResponseDTO.java
@@ -25,8 +25,8 @@ public class LandmarkSaveResponseDTO {
         private String close;
         private Boolean available;
         private String landmarkPhone;
-        private Double lat;
-        private Double lon;
+        private Float lat;
+        private Float lon;
         private String image;
     }
 }
